requirements.txt
fastapi
sqlalchemy
uvicorn
passlib
bcrypt==4.0.1
python-multipart
python-jose